Refactor room controller actions and routes
